CSSTUDIO-2023: cleanup code and structure, move e2e tests to separate workflow
