+++ .github/workflows/e2e.yml
@@ -1,13 +1,13 @@
-name: Test React Application
+name: E2E Tests
 
 on:
-  push:
-    branches: [ "master" ]
-  pull_request:
+  # push:
+  #   branches: [ "master" ]
+  # pull_request:
   workflow_dispatch:
     
 jobs:
-  test:
+  e2e-tests:
     timeout-minutes: 30
     runs-on: ubuntu-latest
 
@@ -22,10 +22,8 @@ jobs:
       with:
         node-version: ${{ matrix.node-version }}
         cache: 'npm'
-    - name: Install Dependencies and Build
-      run: ./scripts/install-dependencies-and-build.sh
-    - name: Run Unit and Integration Tests
-      run: ./scripts/test.sh
+    - name: Run End-To-End Tests
+      run: ./scripts/e2e-test.sh
     - name: Publish Test Results
       if: always()
       uses: actions/upload-artifact@v3
