+++ src/components/Properties/index.js
@@ -1,0 +1,5 @@
+import Properties from "./Properties";
+import { getLogEntryGroupId } from "./utils";
+
+export { Properties, getLogEntryGroupId };
+export default Properties;
