+++ src/components/log/CreateLog/CreateLog.js
@@ -3,8 +3,8 @@ import { useForm } from "react-hook-form";
 import EntryEditor from "../EntryEditor";
 import useFormPersist from "react-hook-form-persist";
 import { useNavigate } from "react-router-dom";
-import ologService, { checkSession, ologServiceWithRetry } from "api/olog-service";
-import { ologClientInfoHeader } from "utils";
+import ologAxiosApi, { checkSession, ologAxiosApiWithRetry } from "api/axios-olog-service";
+import { ologClientInfoHeader } from "api/ologApi";
 import LoadingOverlay from "components/shared/LoadingOverlay";
 
 const CreateLog = ({isAuthenticated}) => {
@@ -69,10 +69,10 @@ const CreateLog = ({isAuthenticated}) => {
                     
                     let url = `/logs/multipart?markup=commonmark`;
                     // Upload the full monty, i.e. log entry and all attachment files, in one single request.
-                    ologService.put(url, multipartFormData, { withCredentials: true, headers: requestHeaders})
+                    ologAxiosApi.put(url, multipartFormData, { withCredentials: true, headers: requestHeaders})
                         .then(async res => {
                             // Wait until the new log entry is available in the search results
-                            await ologServiceWithRetry({
+                            await ologAxiosApiWithRetry({
                                 method: 'GET',
                                 path: `/logs/search?title=${res.data.title}&end=now`,
                                 retries: 5,
