+++ src/components/log/EditLog/EditLog.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useForm } from "react-hook-form";
 import EntryEditor from "../EntryEditor";
-import { ologClientInfoHeader } from "utils";
-import ologService, { ologServiceWithRetry } from "api/olog-service";
+import { ologClientInfoHeader } from "api/ologApi";
+import ologAxiosApi, { ologAxiosApiWithRetry } from "api/axios-olog-service";
 import { useNavigate } from "react-router-dom";
 import LoadingOverlay from "components/shared/LoadingOverlay";
 
@@ -33,7 +33,7 @@ const EditLog = ({log, isAuthenticated}) => {
         
         let url = `/logs/${log.id}?markup=commonmark`;
         // Upload the full monty, i.e. log entry and all attachment files, in one single request.
-        ologService.post(
+        ologAxiosApi.post(
             url, 
             {
                 id: log.id,
@@ -48,7 +48,7 @@ const EditLog = ({log, isAuthenticated}) => {
             )
             .then(async res => {
                 // Wait until the new log entry is available in the search results
-                await ologServiceWithRetry({
+                await ologAxiosApiWithRetry({
                     method: 'GET',
                     path: `/logs/search?title=${res.data.title}&end=now`,
                     retries: 5,
