+++ src/config/localization.js
@@ -15,11 +15,39 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  */
-
+import moment from 'moment';
 import { AdapterMoment } from "@mui/x-date-pickers/AdapterMoment";
 import { LocalizationProvider as MuiLocalizationProvider } from '@mui/x-date-pickers';
 
-const LocalizationProvider = ({children, ...props}) => {
+const shortTimeFormat = 'HH:mm';
+const shortDateFormat = 'YYYY-MM-DD';
+const fullDateTime = 'YYYY-MM-DD HH:mm:ss';
+
+export { moment };
+
+export function formatShortTime(date){
+    return moment(date).format(shortTimeFormat);
+}
+
+export function formatShortDate(date){
+    return moment(date).format(shortDateFormat);
+}
+
+export function formatFullDateTime(date){
+    return moment(date).format(fullDateTime);
+}
+
+/**
+ * Converts a JavaScript Date object to a string on format yyyy-MM-dd HH:mm:ss.
+ */
+export function dateToString(value){
+    return value.getFullYear() + '-' + ('0' + (value.getMonth() + 1)).slice(-2) + '-' +
+           ("0" + value.getDate()).slice(-2) + ' ' + ('0' + value.getHours()).slice(-2) +
+           ':' + ('0' + value.getMinutes()).slice(-2) + ':' +
+           ('0' + value.getSeconds()).slice(-2);
+}
+
+export const LocalizationProvider = ({children, ...props}) => {
 
     return (
         <MuiLocalizationProvider dateAdapter={AdapterMoment} {...props} >
@@ -27,5 +55,3 @@ const LocalizationProvider = ({children, ...props}) => {
         </MuiLocalizationProvider>
     )
 }
-
-export default LocalizationProvider;
