--- src/constants/index.js
@@ -1,1 +1,0 @@
-export const APP_BASE_URL = process.env.REACT_APP_BASE_URL; // e.g. http://localhost:8080/Olog
