+++ src/features/store.js
@@ -1,7 +1,7 @@
 import { configureStore } from "@reduxjs/toolkit";
 import searchPageParamsReducer from "features/searchPageParamsReducer";
 import searchParamsReducer from "features/searchParamsReducer";
-import { ologApi } from "services/ologApi";
+import { ologApi } from "api/ologApi";
 import currentLogEntryReducer from "features/currentLogEntryReducer";
 
 export const setupStore = (preloadedState) => {
