+++ src/hooks/searchParams.test.js
@@ -16,7 +16,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  */
 
-import {queryStringToSearchParameters, searchParamsToQueryString} from './searchParams';
+import {queryStringToSearchParameters, searchParamsToQueryString} from './useSanitizedSearchParams';
 
 const buildTags = (names) => {
     return names.map(name => ({
