+++ src/index.js
@@ -19,14 +19,14 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client';
 import App from 'components/App';
-import { store } from './stores';
+import { store } from './features/store';
 import { Provider } from 'react-redux';
 import { ThemeProvider as StyledComponentsThemeProvider } from 'styled-components';
 import theme, { styledComponentsTheme } from 'config/theme';
 import GlobalStyle from 'config/GlobalStyle';
 import { BrowserRouter } from 'react-router-dom';
 import { ThemeProvider } from '@mui/material';
-import LocalizationProvider from 'config/LocalizationProvider';
+import { LocalizationProvider } from 'config/localization';
 
 const container = document.getElementById('root');
 const root = createRoot(container);
