+++ src/test-utils/index.js
@@ -19,7 +19,7 @@
 import { server } from "mocks/server";
 import { rest } from "msw";
 import { render, within } from "@testing-library/react";
-import { setupStore } from "../stores";
+import { setupStore } from "../features/store";
 import { resultList, testEntry } from "../mocks/fixtures/generators";
 import { AppWrapper } from "./wrappers";
 import { theme, styledComponentsTheme } from "../config/theme";
